+++ packages/components/src/primitives/horizontal-rule/index.native.js
@@ -1,0 +1,16 @@
+/**
+ * External dependencies
+ */
+import Hr from 'react-native-hr';
+
+export const HorizontalRule = ( props ) => {
+	const lineStyle = {
+		backgroundColor: '#555d66',
+		height: 2,
+		...props.lineStyle,
+	};
+
+	return (
+		<Hr { ...props } lineStyle={ lineStyle } />
+	);
+};